week-13/requirements.txt
# ──────────────────────────────────────────────
# Fashion Outfit Generator — Dependencies
# ──────────────────────────────────────────────
# Core ML
torch
torchvision

# Diffusion pipeline
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0

# ControlNet preprocessing
controlnet_aux>=0.0.7

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Evaluation
scikit-image>=0.21.0

# UI
streamlit>=1.30.0
pandas>=2.0.0

# Utilities
safetensors>=0.4.0
